+++ src/main/java/com/jalasoft/selenium/daniel/jauregui/movies/Movie.java
@@ -1,4 +1,4 @@
-package com.jalasoft.selenium.movies;
+package com.jalasoft.selenium.daniel.jauregui.movies;
 
 /***********
  DEPRECATED
