+++ src/main/java/com/jalasoft/selenium/daniel/jauregui/movies/NewRelease.java
@@ -1,4 +1,4 @@
-package com.jalasoft.selenium.movies;
+package com.jalasoft.selenium.daniel.jauregui.movies;
 
 public class NewRelease extends aMovie {
