+++ src/main/java/com/jalasoft/selenium/daniel/jauregui/movies/iMovie.java
@@ -1,4 +1,4 @@
-package com.jalasoft.selenium.movies;
+package com.jalasoft.selenium.daniel.jauregui.movies;
 
 interface iMovie {
     public double calculateRate(int rentedDays, double thisAmount);
