+++ src/main/java/com/jalasoft/selenium/daniel/jauregui/shapes/Main.java
@@ -3,7 +3,7 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package com.jalasoft.selenium.shapes;
+package com.jalasoft.selenium.daniel.jauregui.shapes;
 
 /**
  *
