+++ src/main/java/com/jalasoft/selenium/daniel/jauregui/shapes/Shape.java
@@ -1,4 +1,4 @@
-package com.jalasoft.selenium.shapes;
+package com.jalasoft.selenium.daniel.jauregui.shapes;
 
 public abstract class Shape implements iShape {
 // podria entrar alguna implementacion para heredar
