+++ src/main/java/com/jalasoft/selenium/daniel/jauregui/shapes/iShape.java
@@ -1,4 +1,4 @@
-package com.jalasoft.selenium.shapes;
+package com.jalasoft.selenium.daniel.jauregui.shapes;
 
 
 interface iShape {
