+++ src/test/java/com/jalasoft/selenium/daniel/jauregui/movies/CustomerTest.java
@@ -3,8 +3,13 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package com.jalasoft.selenium.movies;
+package com.jalasoft.selenium.daniel.jauregui.movies;
 
+import com.jalasoft.selenium.daniel.jauregui.movies.Rental;
+import com.jalasoft.selenium.daniel.jauregui.movies.Children;
+import com.jalasoft.selenium.daniel.jauregui.movies.Customer;
+import com.jalasoft.selenium.daniel.jauregui.movies.NewRelease;
+import com.jalasoft.selenium.daniel.jauregui.movies.Regular;
 import org.junit.Before;
 import org.junit.Test;
 import static org.junit.Assert.*;
