+++ src/test/java/com/jalasoft/selenium/daniel/jauregui/movies/NewReleaseTest.java
@@ -1,5 +1,7 @@
-package com.jalasoft.selenium.movies;
+package com.jalasoft.selenium.daniel.jauregui.movies;
 
+import com.jalasoft.selenium.daniel.jauregui.movies.aMovie;
+import com.jalasoft.selenium.daniel.jauregui.movies.NewRelease;
 import org.junit.Before;
 import org.junit.Test;
 import static org.junit.Assert.*;
@@ -8,62 +10,66 @@ import static org.junit.Assert.*;
  *
  * @author Daniel Jauregui
  */
-public class RegularTest {
+public class NewReleaseTest {
     
     private aMovie instance;
     private final double thisAmount = 0.0;
     private final double DELTA = 0.0;
-    
+
     @Before
     public void setUp() {
         // Given
-        instance = new Regular("Avatar",0);
+        instance = new NewRelease("Dr Strange",0);
     }
     
     /**
-     * Test of calculateRate method, of class Regular.
+     * Test of calculateRate method, of class NewRelease.
      */
     @Test
-    public void testCalculateRateWithRentedDaysMajorThan2() {
-        //When
-        int rentedDays = 12;
-        double expResult = 17;
-        double result = instance.calculateRate(rentedDays, thisAmount);
-        //Then
-        System.out.println("testCalculateRateWithRentedDaysMajorThan2");
-        assertEquals(expResult, result, DELTA);
-    }
-
-    @Test
-    public void testCalculateRateWithRentedDaysEqualTo2() {
-        //When
-        int rentedDays = 2;
-        double expResult = 2;
-        double result = instance.calculateRate(rentedDays, thisAmount);
-        //Then
-        System.out.println("testCalculateRateWithRentedDaysEqualTo2");
-        assertEquals(expResult, result, DELTA);
-    }
-
-    @Test
-    public void testCalculateRateWithRentedDaysLessThan2() {
+    public void testCalculateRateWithRentedDays1Day() {
         //When
         int rentedDays = 1;
-        double expResult = 2;
+        double expResult = 3;
         double result = instance.calculateRate(rentedDays, thisAmount);
         //Then
-        System.out.println("testCalculateRateWithRentedDaysLessThan2");
+        System.out.println("testCalculateRateWithRentedDays1Day");
         assertEquals(expResult, result, DELTA);
+        
     }
     
     @Test
-    public void testCalculateRateWithRentedDaysEqualTo0() {
+    public void testCalculateRateWithRentedDays1Week() {
+        //When
+        int rentedDays = 7;
+        double expResult = 21;
+        double result = instance.calculateRate(rentedDays, thisAmount);
+        //Then
+        System.out.println("testCalculateRateWithRentedDays1Week");
+        assertEquals(expResult, result, DELTA);
+        
+    }
+    
+    @Test
+    public void testCalculateRateWithRentedDays1Month() {
+        //When
+        int rentedDays = 30;
+        double expResult = 90;
+        double result = instance.calculateRate(rentedDays, thisAmount);
+        //Then
+        System.out.println("testCalculateRateWithRentedDays1Month");
+        assertEquals(expResult, result, DELTA);
+        
+    }
+    
+    @Test
+    public void testCalculateRateWithRentedDays0Days() {
         //When
         int rentedDays = 0;
         double expResult = 0;
         double result = instance.calculateRate(rentedDays, thisAmount);
         //Then
-        System.out.println("testCalculateRateWithRentedDaysEqualTo0");
+        System.out.println("testCalculateRateWithRentedDays0d");
         assertEquals(expResult, result, DELTA);
+        
     }
 }
