+++ src/test/java/com/jalasoft/selenium/daniel/jauregui/movies/RegularTest.java
@@ -1,5 +1,7 @@
-package com.jalasoft.selenium.movies;
+package com.jalasoft.selenium.daniel.jauregui.movies;
 
+import com.jalasoft.selenium.daniel.jauregui.movies.aMovie;
+import com.jalasoft.selenium.daniel.jauregui.movies.Regular;
 import org.junit.Before;
 import org.junit.Test;
 import static org.junit.Assert.*;
@@ -8,66 +10,62 @@ import static org.junit.Assert.*;
  *
  * @author Daniel Jauregui
  */
-public class NewReleaseTest {
+public class RegularTest {
     
     private aMovie instance;
     private final double thisAmount = 0.0;
     private final double DELTA = 0.0;
-
+    
     @Before
     public void setUp() {
         // Given
-        instance = new NewRelease("Dr Strange",0);
+        instance = new Regular("Avatar",0);
     }
     
     /**
-     * Test of calculateRate method, of class NewRelease.
+     * Test of calculateRate method, of class Regular.
      */
     @Test
-    public void testCalculateRateWithRentedDays1Day() {
+    public void testCalculateRateWithRentedDaysMajorThan2() {
+        //When
+        int rentedDays = 12;
+        double expResult = 17;
+        double result = instance.calculateRate(rentedDays, thisAmount);
+        //Then
+        System.out.println("testCalculateRateWithRentedDaysMajorThan2");
+        assertEquals(expResult, result, DELTA);
+    }
+
+    @Test
+    public void testCalculateRateWithRentedDaysEqualTo2() {
+        //When
+        int rentedDays = 2;
+        double expResult = 2;
+        double result = instance.calculateRate(rentedDays, thisAmount);
+        //Then
+        System.out.println("testCalculateRateWithRentedDaysEqualTo2");
+        assertEquals(expResult, result, DELTA);
+    }
+
+    @Test
+    public void testCalculateRateWithRentedDaysLessThan2() {
         //When
         int rentedDays = 1;
-        double expResult = 3;
+        double expResult = 2;
         double result = instance.calculateRate(rentedDays, thisAmount);
         //Then
-        System.out.println("testCalculateRateWithRentedDays1Day");
+        System.out.println("testCalculateRateWithRentedDaysLessThan2");
         assertEquals(expResult, result, DELTA);
-        
     }
     
     @Test
-    public void testCalculateRateWithRentedDays1Week() {
-        //When
-        int rentedDays = 7;
-        double expResult = 21;
-        double result = instance.calculateRate(rentedDays, thisAmount);
-        //Then
-        System.out.println("testCalculateRateWithRentedDays1Week");
-        assertEquals(expResult, result, DELTA);
-        
-    }
-    
-    @Test
-    public void testCalculateRateWithRentedDays1Month() {
-        //When
-        int rentedDays = 30;
-        double expResult = 90;
-        double result = instance.calculateRate(rentedDays, thisAmount);
-        //Then
-        System.out.println("testCalculateRateWithRentedDays1Month");
-        assertEquals(expResult, result, DELTA);
-        
-    }
-    
-    @Test
-    public void testCalculateRateWithRentedDays0Days() {
+    public void testCalculateRateWithRentedDaysEqualTo0() {
         //When
         int rentedDays = 0;
         double expResult = 0;
         double result = instance.calculateRate(rentedDays, thisAmount);
         //Then
-        System.out.println("testCalculateRateWithRentedDays0d");
+        System.out.println("testCalculateRateWithRentedDaysEqualTo0");
         assertEquals(expResult, result, DELTA);
-        
     }
 }
